# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storytail API",
        "description": "API documentation for Storytail",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/popular-books": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Livros mais populares",
                "description": "Retorna os 3 livros mais clicados nos últimos 3 meses",
                "operationId": "0e836a1c2fb8f650c0d4fe8fb27053a2",
                "responses": {
                    "200": {
                        "description": "Lista dos livros mais populares",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "clicks_count": {
                                                "type": "integer",
                                                "example": 150
                                            },
                                            "book": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "O Livro Mais Popular"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "read_time": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "access_level": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "authors": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar pedido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/peak-usage-times": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Horários de pico",
                "description": "Retorna o horário com mais cliques em livros",
                "operationId": "fc81d60e4bf17e1e67311873c67bbcc5",
                "responses": {
                    "200": {
                        "description": "Horário com mais acessos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "hour": {
                                                "description": "Hora do dia (0-23)",
                                                "type": "integer",
                                                "example": 14
                                            },
                                            "clicks_count": {
                                                "description": "Número de cliques nessa hora",
                                                "type": "integer",
                                                "example": 250
                                            },
                                            "percentage": {
                                                "description": "Percentagem do total de cliques",
                                                "type": "number",
                                                "example": 15.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar pedido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}